# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: HustleCoin V2 API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: List all active tasks with the current user's status
      responses:
        "200":
          description: Task list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TasksListResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tasks/{id}:
    get:
      operationId: getTaskById
      tags: [tasks]
      summary: Get a single task with the current user's status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskItem"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tasks/{id}/start:
    post:
      operationId: startTask
      tags: [tasks]
      summary: Mark a task as in progress
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskItem"
        "400":
          description: Task already completed or pending
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tasks/{id}/complete:
    post:
      operationId: completeTask
      tags: [tasks]
      summary: Complete a task. Automatic tasks grant reward immediately; manual tasks go to pending approval.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task completion result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskActionResult"
        "400":
          description: Already completed, pending, or condition not met
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tasks/{id}/claim:
    post:
      operationId: claimTaskReward
      tags: [tasks]
      summary: Claim reward for a manually approved task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reward claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskActionResult"
        "400":
          description: Nothing to claim or already claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    TaskItem:
      type: object
      required: [id, title, description, reward, taskType, category, status, canClaim]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        reward:
          type: integer
        link:
          type: ["string", "null"]
        taskType:
          type: string
          enum: [automatic, manual]
        category:
          type: string
          enum: [daily, one_time]
        status:
          type: string
          enum: [available, in_progress, pending_approval, completed]
        canClaim:
          type: boolean
        lastCompletedAt:
          type: ["string", "null"]
    TasksListResponse:
      type: object
      required: [tasks]
      properties:
        tasks:
          type: array
          items:
            $ref: "#/components/schemas/TaskItem"
    TaskActionResult:
      type: object
      required: [status, reward, newBalance]
      properties:
        status:
          type: string
          enum: [completed, pending_approval]
        reward:
          type: integer
        newBalance:
          type: integer
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...